fix(suite): simplify and fix  op_return as receive address
