+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/TxDetailModal/Detail/AdvancedTxDetails/IODetails/CollapsibleIOSection.tsx
@@ -1,0 +1,36 @@
+import { ReactNode } from 'react';
+
+import { CollapsibleBox } from '@trezor/components';
+
+import { IOGroup, IOGroupProps } from './IOGroup';
+
+type CollapsibleIOSectionProps = IOGroupProps & {
+    heading?: ReactNode;
+    opened?: boolean;
+};
+
+export const CollapsibleIOSection = ({
+    tx,
+    inputs,
+    outputs,
+    heading,
+    opened,
+    isPhishingTransaction,
+}: CollapsibleIOSectionProps) =>
+    inputs?.length || outputs?.length ? (
+        <CollapsibleBox
+            heading={heading}
+            defaultIsOpen={opened}
+            paddingType="none"
+            fillType="none"
+            hasDivider={false}
+        >
+            <IOGroup
+                tx={tx}
+                inputs={inputs}
+                outputs={outputs}
+                isPhishingTransaction={isPhishingTransaction}
+                isUtxoBased
+            />
+        </CollapsibleBox>
+    ) : null;
